Move testdata protos under internal/ and use more compatible paths.
