+++ internal/testdata/test_proto_gogo/test_proto_gogo.proto
@@ -16,10 +16,19 @@
 
 syntax = "proto2";
 
-option go_package = "github.com/stripe/skycfg/test_proto";
 package skycfg.test_proto;
 
-message MessageV2 {
+import "github.com/gogo/protobuf/gogoproto/gogo.proto";
+import "github.com/stripe/skycfg/internal/testdata/test_proto/test_proto_v2.proto";
+import "google/protobuf/duration.proto";
+
+option go_package = "github.com/stripe/skycfg/internal/testdata/test_proto_gogo";
+
+message MessageGogo {
+  option (gogoproto.sizer) = false;
+  option (gogoproto.marshaler) = false;
+  option (gogoproto.unmarshaler) = false;
+
   optional int32  f_int32   = 1;
   optional int64  f_int64   = 2;
   optional uint32 f_uint32  = 3;
@@ -29,10 +38,14 @@ message MessageV2 {
   optional string f_string  = 7 [default="default_str"];;
   optional bool   f_bool    = 8;
 
-  optional MessageV2 f_submsg = 9;
+  optional MessageV2 f_submsg = 9 [
+    (gogoproto.nullable) = false
+  ];
 
   repeated string        r_string = 10;
-  repeated MessageV2     r_submsg = 11;
+  repeated MessageV2     r_submsg = 11 [
+    (gogoproto.nullable) = false
+  ];
 
   map<string, string>    map_string = 12;
   map<string, MessageV2> map_submsg = 13;
@@ -56,10 +69,10 @@ message MessageV2 {
 
   optional bytes f_bytes = 19;
 
-  // NEXT: 20
-}
+  optional google.protobuf.Duration f_duration = 20 [
+    (gogoproto.stdduration) = true,
+    (gogoproto.nullable) = false
+  ];
 
-enum ToplevelEnumV2 {
-  TOPLEVEL_ENUM_V2_A = 0;
-  TOPLEVEL_ENUM_V2_B = 1;
+  // NEXT: 21
 }
